pyproject.toml
[project]
name = "meal-planner"
version = "0.1.0"
description = "AI meal planner"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "kagglehub",
  "langchain-ollama",
  "langgraph",
  "pandas",
  "pgvector",
  "psycopg[binary]",
  "pydantic",
  "python-dotenv",
  "tqdm",
  "typer",
]

[project.scripts]
meal-planner = "meal_planner.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
